requirements.txt
streamlit==1.38.0
openai==1.35.10
python-dotenv==1.0.1
pillow==10.4.0
git+https://github.com/huggingface/transformers.git
torch
torchvision
# llama-index-core==0.11.21
# pydantic==2.9.2